+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 - New SMAC logo
 - Fix doc link in README
 
+## Documentation
+- Ask and tell without initial design and warmstarting
+
+## Bugfixes 
+- Ask and tell without initial design may still return a config from the initial design.  
+
 # 2.3.0
 
 ## Documentation