+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Add target function with additional arguments (#1134)
 
 
+## Improvements
+- `target_function` becomes optional in Facade when using ask and tell exclusively (#946)
+
 # 2.3.0
 
 ## Documentation